DropOffUI Implemented
Added a simple bubble icon that prompts the player for the requirements to "refine" specific ingredients. Currently works for just the boiler but can be added to other dropoffzone objects with a simple image icon assignment in the inspector. Added an IconScreenshot script to grab snapshots of the 3D models in the samplescene