#fm-26 fixed inventory asset on shop system
Dictionaries are not serializable by default and therefore do not populate in the scriptable object asset.

Have to mirror a dictionary in a list to get it to show up. Will probably want to scrap the dictionary and just use a list to simplify things in the future but proceeding as is for now.

We've progressed to the point where the shopper finds an item they want that's in stock and we get an error when they try to get the price. Expecting the itemKey (FarmShop.cs:154) is incorrect.